Return a local table variable for the main module
The codebase has been modified in order for the main module and its
submodules to return local variables, so as not to pollute the global
environment.  Also, now the returned table, only contains the public API
methods, as the submodules and their methods can be accessed without
having them stored there.  The documentation, examples and tests have
been also updated to reflect this change.